Move runtime and event loop ext to their own files
